Fix mobile sdk - register wallet
